refactor(settings): extract about/support/shortcuts/storage/beta/models/privacy pages
Continues the TranscriptedSettingsView decomposition. Also fixes the repo
command contract tests: several `readRepoTextFile("...TranscriptedSettingsView.swift")`
call sites were asserting on page content that has now moved to sibling
extension files. Introduced readSettingsSurfaceContents() to concatenate the
main file with its TranscriptedSettingsView+*.swift siblings so these
contract tests keep working regardless of which physical file a page's code
lives in, instead of re-patching them on every future page move.
