+++ Sources/UI/Settings/TranscriptedSettingsView+Support.swift
@@ -1,0 +1,62 @@
+import SwiftUI
+
+extension TranscriptedSettingsView {
+    var supportPage: some View {
+        VStack(alignment: .leading, spacing: 20) {
+            SettingsPageIntro(
+                title: "Support",
+                summary: "Need help, found a bug, or want to send feedback? Email is the best way to reach the team building Transcripted."
+            )
+
+            SupportActionCard(
+                symbolName: "envelope.fill",
+                title: "Email support",
+                detail: "Send feedback, ask for help, or tell us what felt broken. This opens a prefilled email to help@transcripted.app.",
+                buttonTitle: "Email support",
+                buttonSymbolName: "paperplane.fill",
+                tone: .primary,
+                status: nil,
+                isEnabled: true
+            ) {
+                trackSettingsAction("submit_feedback", page: .support)
+                actions.sendFeedback()
+            }
+
+            SupportActionCard(
+                symbolName: "waveform.path.ecg",
+                title: "Send diagnostics",
+                detail: "Had an error or something felt broken? Send a privacy-safe diagnostic event so we can investigate and try to fix it.",
+                buttonTitle: "One-click send diagnostics",
+                buttonSymbolName: "bolt.fill",
+                tone: .secondary,
+                status: diagnosticsActionStatus,
+                isEnabled: CrashReporter.isAvailable && crashReportingEnabled
+            ) {
+                trackSettingsAction("send_diagnostic_event", page: .support)
+                sendDiagnosticEvent()
+            }
+
+            SupportPrivacyNote()
+        }
+        .accessibilityIdentifier("transcripted.settings.page.support")
+    }
+
+    private func sendDiagnosticEvent() {
+        guard CrashReporter.isAvailable else {
+            diagnosticsActionStatus = "Sentry is not configured in this build yet."
+            return
+        }
+
+        guard crashReportingEnabled else {
+            diagnosticsActionStatus = "Turn on crash and error reports first."
+            return
+        }
+
+        guard let eventID = actions.sendDiagnosticEvent() else {
+            diagnosticsActionStatus = "Diagnostic event could not be queued."
+            return
+        }
+
+        diagnosticsActionStatus = "Queued diagnostic event \(eventID.prefix(8))."
+    }
+}
